pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-knowledge-base"
version = "0.1.0"
description = "Local-first LLM knowledge base builder for Obsidian-style markdown vaults."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openai-agents",
  "typer>=0.12",
  "rich>=13.7",
  "pypdf>=5.0.0",
  "matplotlib>=3.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
]

[project.scripts]
lkb = "llm_kb.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

